chatbot/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
attrs==25.3.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.7.9
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
dataclasses-json==0.6.7
    # via langchain-community
faiss-cpu==1.11.0
    # via -r requirements.in
fastapi==0.116.1
    # via -r requirements.in
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via huggingface-hub
google-ai-generativelanguage==0.6.4
    # via google-generativeai
google-api-core[grpc]==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.176.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.5.4
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.3
    # via sqlalchemy
grpcio==1.73.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via langsmith
httpx-sse==0.4.1
    # via langchain-community
huggingface-hub==0.33.4
    # via
    #   langchain-huggingface
    #   tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.26
    # via langchain-community
langchain-community==0.3.27
    # via -r requirements.in
langchain-core==0.3.68
    # via
    #   langchain
    #   langchain-community
    #   langchain-huggingface
    #   langchain-text-splitters
langchain-huggingface==0.3.0
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.4.5
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.1
    # via
    #   faiss-cpu
    #   langchain-community
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==4.25.8
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   google-generativeai
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via langchain-community
pyparsing==3.2.3
    # via httplib2
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
requests==2.32.4
    # via
    #   google-api-core
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via
    #   langchain
    #   langchain-community
starlette==0.47.1
    # via fastapi
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tokenizers==0.21.2
    # via langchain-huggingface
tqdm==4.67.1
    # via
    #   google-generativeai
    #   huggingface-hub
typing-extensions==4.14.1
    # via
    #   fastapi
    #   google-generativeai
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
yarl==1.20.1
    # via aiohttp
zstandard==0.23.0
    # via langsmith
